# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SewoApp API Documentation",
        "description": "Dokumentasi API untuk SewoApp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/order-details": {
            "get": {
                "tags": [
                    "Order Details"
                ],
                "summary": "Mendapatkan daftar semua detail pesanan",
                "operationId": "e7b43d41efab403b7ce46f3a475d1d05",
                "responses": {
                    "200": {
                        "description": "Daftar detail pesanan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderDetail"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Order Details"
                ],
                "summary": "Membuat detail pesanan baru",
                "operationId": "dd93a1f6bc210c23cd827f3bbfe0975d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_order",
                                    "id_vehicle",
                                    "price",
                                    "total_days",
                                    "total_price"
                                ],
                                "properties": {
                                    "id_order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_vehicle": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250000
                                    },
                                    "total_days": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1250000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detail pesanan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/order-details/{id}": {
            "get": {
                "tags": [
                    "Order Details"
                ],
                "summary": "Mendapatkan detail pesanan berdasarkan ID",
                "operationId": "518b6a2d6a16d8d275e5a5f13ee77fb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID detail pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pesanan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Detail pesanan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Order Details"
                ],
                "summary": "Mengupdate detail pesanan",
                "operationId": "322a821a2c6e7aeaf6befeb20760c583",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID detail pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_vehicle": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250000
                                    },
                                    "total_days": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1250000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detail pesanan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Detail pesanan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order Details"
                ],
                "summary": "Menghapus detail pesanan",
                "operationId": "d4db6edcff565efa58eb69b176062899",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID detail pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Detail pesanan berhasil dihapus"
                    },
                    "404": {
                        "description": "Detail pesanan tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderDetail": {
                "required": [
                    "id_order",
                    "id_vehicle",
                    "price",
                    "total_days",
                    "total_price"
                ],
                "properties": {
                    "id_order_detail": {
                        "description": "ID detail pesanan",
                        "type": "integer",
                        "example": 1
                    },
                    "id_order": {
                        "description": "ID pesanan",
                        "type": "integer",
                        "example": 1
                    },
                    "id_vehicle": {
                        "description": "ID kendaraan",
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "description": "Harga per hari",
                        "type": "number",
                        "format": "float",
                        "example": 250000
                    },
                    "total_days": {
                        "description": "Total hari sewa",
                        "type": "integer",
                        "example": 5
                    },
                    "total_price": {
                        "description": "Total harga",
                        "type": "number",
                        "format": "float",
                        "example": 1250000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}